% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cycle_model}
\alias{cycle_model}
\title{The hidden semi-Markov model for identifying cycles from bleeding data.}
\format{
An object of class \code{hsmm} of length 10.
}
\usage{
cycle_model
}
\description{
A hidden semi-Markov model from the \code{HiddenSemiMarkov} package.
}
\keyword{datasets}
