% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formating_functions.R
\name{.compute_phase_and_cycleday}
\alias{.compute_phase_and_cycleday}
\title{Compute cycledays and phases from dates and provided starts of cycles.}
\usage{
.compute_phase_and_cycleday(data, sep_event)
}
\arguments{
\item{data}{a data.frame that contains symptoms reported subjects. The data must be in a long format and have the following columns: \code{SUBJECT}, \code{ITEM}, \code{DRSP_score}}

\item{sep_event}{a character XXX}
}
\value{
the same data.frame as data with additional columns.
}
\description{
Compute cycledays and phases from dates and provided starts of cycles.
}
\examples{
random_data =
    expand.grid(
        SUBJECT = 1,
        ITEM = 1:24,
        date = seq(as.Date("2020-01-01"), as.Date("2020-03-01"), by = 1)
        )
random_data$DRSP_score = sample(1:6, nrow(random_data), replace = TRUE)
random_data$is_cycle_start = ifelse(random_data$date \%in\% c(as.Date("2020-01-15"), as.Date("2020-02-20")), TRUE, FALSE)
augmented_data = .compute_phase_and_cycleday(random_data, sep_event = "ovulation")
colnames(augmented_data)
}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
