% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dsm5_dict}
\alias{dsm5_dict}
\title{PMDD DSM-5 items and their domains.}
\format{
A data frame with 24 rows and 5 variables:
\describe{
  \item{DRSP}{item number}
  \item{DRSP_full_code}{item number with its names}
  \item{DRSP_desc}{item description}
  \item{DSM5_SYMPTOM_DOMAIN}{item domain, as defined in the DSM-5}
  \item{SYMPTOM_CATEGORY}{item category, as defined in the DSM-5. Can be a "Core Emotional Symptom", a "Secondary Symptom" or an "Interference".}
}
}
\usage{
dsm5_dict
}
\description{
A dictionary listing the 24 DSM-5 items for which subjects can report a DRSP (Daily Record of Severity of Problems) score
}
\keyword{datasets}
