% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formating_functions.R
\name{is_cpass_data}
\alias{is_cpass_data}
\title{Checks if data is a "CPASS" data.frame.}
\usage{
is_cpass_data(data, silent = FALSE)
}
\arguments{
\item{data}{a data.frame that contains symptoms reported subjects. The data must be in a long format and have the following columns: \code{SUBJECT}, \code{CYCLE}, \code{DAY}, \code{DRSP}, \code{DRSP_score}}

\item{silent}{(optional) logical, default = \code{FALSE}. Should warning messages be printed?}
}
\value{
a logical.
}
\description{
Checks if data is a "CPASS" data.frame.
}
\examples{
random_data = expand.grid(SUBJECT = 1, CYCLE = 1:2, DAY = c(1:10,-10:-1), ITEM = 1:24)
random_data$DRSP_score = sample(1:6, nrow(random_data), replace = TRUE)
cpass_data = as_cpass_data(random_data, sep_event = "menses")
colnames(cpass_data)
is_cpass_data(random_data)
is_cpass_data(cpass_data)
}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
