% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_subject_cycle_obs}
\alias{plot_subject_cycle_obs}
\title{Visualize the scores of a subject's cycle}
\usage{
plot_subject_cycle_obs(data = data.frame(), add_diagnosis = TRUE)
}
\arguments{
\item{data}{a data.frame that contains the symptoms reported in ONE cycle by ONE subject. The data must be in a long format and have the following columns: \code{SUBJECT}, \code{CYCLE}, \code{DAY}, \code{DRSP}, \code{score}}

\item{add_diagnosis}{logical. If \code{TRUE} (default), the diagnoses at the DRSP, DSM5-DOMAIN and CYCLE levels are displayed together with the subject's reported scores. If \code{FALSE}, only the reported scores are displayed.}
}
\value{
a ggplot object
}
\description{
This function visualizes the scores (raw data) of a single subject's cycle. It also internally calls the \code{CPASS} function to provides a visualization of the diagnosis at the DRSP, DSM5 DOMAIN and CYCLE level.
}
\examples{
data(PMDD_data)
input = PMDD_data \%>\% filter(SUBJECT == 2, CYCLE == 1)
plot_subject_cycle_obs(data = input)

}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
