% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_subject_data_and_diagnosis}
\alias{plot_subject_data_and_diagnosis}
\title{Visualize all the observations and the diagnosis summary of a single subject.}
\usage{
plot_subject_data_and_diagnosis(
  data = data.frame(),
  save_as_pdf = TRUE,
  pdf_path = "",
  pdf_name = "",
  max_score_color = "tomato"
)
}
\arguments{
\item{data}{a data.frame that contains the symptoms reported by ONE subject.
The data must be in a long format and have the following columns: \code{SUBJECT}, \code{CYCLE}, \code{DAY}, \code{DRSP}, \code{score}}

\item{save_as_pdf}{logical.
If \code{TRUE} (default), this function saves a pdf with the visualizations.
The pdf path and name can be specified with the arguments \code{pdf_path} and \code{pdf_name}.
If \code{FALSE}, this function returns a ggplot object. The same object is also returned, but invisibly, when \code{save_as_pdf = TRUE} (see examples).}

\item{pdf_path}{string. Specifies the path to the folder in which the pdf should be saved.
By default, the path is an empty string so the pdf is be saved in the current working directory.}

\item{pdf_name}{string. Specifies the name of the pdf.
By default, the name of the pdf is \code{CPASS_SUBJECT_X.pdf}, where \code{X} is replaced by the subject unique identifier.}
}
\value{
a ggplot object if \code{save_as_pdf = FALSE}. The same object is returned invisibly by default (\code{save_as_pdf = TRUE}).
}
\description{
By default, this function saves the visualization in a pdf. The pdf heigth scales with the number of cycles reported for the subject.
}
\examples{
data(PMDD_data)
input = PMDD_data \%>\% filter(SUBJECT == 2)
p = plot_subject_data_and_diagnosis(data = input)
p
}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
