% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_subject_diagnosis}
\alias{plot_subject_diagnosis}
\title{Visualize the subject diagnosis}
\usage{
plot_subject_diagnosis(data = data.frame())
}
\arguments{
\item{data}{a data.frame that contains the symptoms reported by ONE subject. The data must be in a long format and have the following columns: \code{SUBJECT}, \code{CYCLE}, \code{DAY}, \code{DRSP}, \code{score}}
}
\value{
a ggplot object
}
\description{
This function calls the \code{CPASS} function and provides a visualization of the diagnosis.
}
\examples{
data(PMDD_data)
input = PMDD_data \%>\% filter(SUBJECT == 2)
plot_subject_diagnosis(data = input)

}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
