% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_subject_dx}
\alias{plot_subject_dx}
\title{Visualize the subject diagnosis}
\usage{
plot_subject_dx(
  data = data.frame(),
  color_summary = c("complementary", "rainbow")
)
}
\arguments{
\item{data}{a \code{cpass} data frame
(use \code{as_cpass_data} to convert your data into \code{cpass} data)
that contains the symptoms reported by ONE subject.}

\item{color_summary}{string.
Either \code{"complementary"} (default) or \code{"rainbow"}
specifying the type of color scheme for the ITEM items
in the diagnosis summary.
If \code{"complementary"}, the colors are chosen complementary
within a domain;
if \code{"rainbow"}, the DRSP items colors are different for each item and
chosen from a rainbow palette.}
}
\value{
a ggplot object
}
\description{
This function calls the \code{CPASS} function and
provides a visualization of the diagnosis.
}
\examples{

library(magrittr)
library(dplyr)

data(PMDD_data)
input <-
  PMDD_data \%>\%
  dplyr::filter(subject == 2) \%>\%
  as_cpass_data(., sep_event = "menses")
plot_subject_dx(data = input)
}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
