% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_cpass_data.R
\name{simulate_cpass_data}
\alias{simulate_cpass_data}
\title{Simulates CPASS data}
\usage{
simulate_cpass_data(
  n_subjects = 1,
  n_cycles = 1,
  baseline = 1,
  pmdd_intensity = 0.5,
  n_items = 24,
  missingness_rate = 0
)
}
\arguments{
\item{n_subjects}{(positive integer) The number of subjects for which data
need to be simulated}

\item{n_cycles}{(positive integer(s)) The number of cycles for each subject.
This parameter (as well as all following parameters) can be of length
\code{1} (and all subjects have the same number of cycles) or of length
\code{n_subjects} to determine the number of cycles independently
for each subject.}

\item{baseline}{(positive values in 1-6) Baseline levels for severity of
problems for all subjects (if \code{baseline} is of length \code{1}) or
each subject (if \code{baseline} is of length \code{n_subjects}).}

\item{pmdd_intensity}{change from baseline during the pre-menstrual phase.}

\item{n_items}{(positive integers in 1-24) number of items for which there is
a variation in the pre-menstrual phase. Can be of length \code{1}
(same \code{n_items} for all subjects) or of length \code{n_subjects}
(specific \code{n_items} for each subject).}

\item{missingness_rate}{(positive values in 0-0.9) The fraction of missing
days for all (if of length \code{1}) or each (if of length \code{n_subjects})
subject.}
}
\value{
a \code{data.frame} providing the DRSP scores for each subject,
cycle, cycle day (within the pre- and post- menstrual phase) and item.
}
\description{
This function simulates DRSP data for \code{n_subjects}.
Options allow to choose the number of cycles per subject (\code{n_cycles}),
the increase in symptoms rating (\code{pmdd_intensity}), the baseline levels
severity of problems for subjects (\code{baseline}), the number of items
varying with the menstrual cycle (\code{n_items}), and the missingness rate
(\code{missingness_rate}), which is the fraction of days for which data are
missing.
}
\seealso{
\code{simulate_cycle}
}
