% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formating_functions.R
\name{supported_data_format}
\alias{supported_data_format}
\title{Supported data format for the \code{as_cpass_data} function.}
\usage{
supported_data_format(as_list = FALSE)
}
\arguments{
\item{as_list}{(optional, default = \code{FALSE}) logical specifying if
the supported format should be printed to the console
or returned as a list object.}
}
\value{
a list if \code{as_list = TRUE}.
}
\description{
This function prints (or return) a list of the format currently
supported by the \code{cpass} package.
}
\examples{
supported_data_format()

}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
