% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_formating_functions.R
\name{as_cpass_data}
\alias{as_cpass_data}
\title{Transforms into a "cpass" data.frame.}
\usage{
as_cpass_data(data, sep_event = NULL, silent = FALSE, verbose = TRUE)
}
\arguments{
\item{data}{a data.frame that contains symptoms reported subjects.
The data must be in a long format,
_i.e._ one row per subject, day (date or cycle + cycleday) and DRSP item.
The data.frame must have the following columns:
\code{subject}, \code{item}, \code{drsp_score}
as well as two columns defining time.
These two columns can be
\code{cycle} + \code{day},
\code{date} + \code{is_cycle_start}, or
\code{date} + \code{bleeding}.
See \code{supported_data_format} for details.}

\item{sep_event}{(required) character (\code{"menses"} or \code{"ovulation"})
specifying whether the menstrual event separating the two phases of interest
(the "pre" and "post" weeks) is the menses or ovulation.
If the menses is the separating event,
the week preceding the menstruation ("pre") is compared
to the week following the menstruation ("post").
In that case, the "pre" and the "post" week do NOT belong to the same cycle.
If ovulation is the separating event,
then the two phases belong to the same cycle.}

\item{silent}{(optional) logical, default = \code{FALSE}.
Should warning messages be printed?}

\item{verbose}{(optional) logical.
Should the function print additional information?
(mostly used for debugging purposes)}
}
\value{
a data.frame.
}
\description{
This function checks if the data can be used by the CPASS functions
and transforms it into a "cpass" data.frame.
}
\examples{
random_data <-
   expand.grid(
       subject = 1,
       cycle = 1:2,
       day = c(1:10,-10:-1),
       item = 1:24
       )
random_data$drsp_score <- sample(1:6, nrow(random_data), replace = TRUE)
cpass_data <- as_cpass_data(random_data, sep_event = "menses")
colnames(cpass_data)
}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
