% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPASS_function.R
\name{cpass}
\alias{cpass}
\title{Applies the C-PASS procedure for PMDD and MRMD diagnoses}
\usage{
cpass(data, silent = FALSE)
}
\arguments{
\item{data}{a \code{cpass.data} data frame
that contains the symptom DRSP scores reported by the subjects.
To transform your data into a \code{cpass.data} object,
use the function \code{as_cpass_data()}.}

\item{silent}{a \code{logical}
specifying is the function should print messages or run silently.
Default is \code{FALSE}.}
}
\value{
The cpass function returns a list with 6 elements;
each of them is a table (\code{data.frame}):
\describe{
  \item{\code{subject_level_diagnosis}}{The subject-level diagnoses}
  \item{\code{cycle_level_diagnosis}}{The cycle-level diagnoses}
  \item{\code{DSM5_domains_level_diagnosis}}{The diagnoses at the DSM5 DOMAINS level}
  \item{\code{DRSP_level_diagnosis}}{The diagnoses at the individual DRSP items level}
  \item{\code{daily_summary_DRSP}}{The daily summary of each DRSP for each subject}
  \item{\code{summary_DRSP}}{The percent change in DRSP items between the pre- and post-menstrual phase}
}
}
\description{
This function implements the C-PASS procedure and
returns a list of 6 tables of summaries and
diagnoses at different levels (subjects, cycles, domains, etc).
}
\examples{
data(PMDD_data)
cpass_input <- as_cpass_data(PMDD_data, sep_event = "menses")
output <- cpass(cpass_input)
head(output$subject_level_diagnosis)
}
\keyword{C-PASS}
\keyword{MRMD}
\keyword{PMDD}
\keyword{cpass}
