% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPASS_functions.R
\name{CPASS}
\alias{CPASS}
\title{Applies the C-PASS procedure for PMDD and MRMD diagnoses}
\usage{
CPASS(data)
}
\arguments{
\item{data}{a data.frame that contains the symptoms reported by the subjects. The data must be in a long format and have the following columns: SUBJECT, CYCLE, DAY, DRSP, score}
}
\value{
6 tables: (1) the subject-level diagnoses; (2) the cycle-level diagnoses; (3) the diagnoses at the DSM5 DOMAINS level; (4) the diagnoses at the individual DRSP items level; (5) the daily summary of each DRSP for each subject and (6) the percentage change in DRSP items between the pre- and post-menstrual phase
}
\description{
This function implements the C-PASS procedure and returns.
}
\examples{

}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
