% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization_functions.R
\name{plot_subject_data_and_diagnosis}
\alias{plot_subject_data_and_diagnosis}
\title{Visualize all the observations and the diagnosis summary of a single subject.}
\usage{
plot_subject_data_and_diagnosis(
  data = data.frame(),
  save_as_pdf = TRUE,
  pdf_path = "",
  pdf_name = "",
  color_max_score = "tomato",
  color_summary = c("complementary", "rainbow")
)
}
\arguments{
\item{data}{a data.frame that contains the symptoms reported by ONE subject.
The data must be in a long format and have the following columns: \code{SUBJECT}, \code{CYCLE}, \code{DAY}, \code{ITEM}, \code{DRSP_score}}

\item{save_as_pdf}{logical.
If \code{TRUE} (default), this function saves a pdf with the visualizations.
The pdf path and name can be specified with the arguments \code{pdf_path} and \code{pdf_name}.
If \code{FALSE}, this function returns a ggplot object. The same object is also returned, but invisibly, when \code{save_as_pdf = TRUE} (see examples).}

\item{pdf_path}{string. Specifies the path to the folder in which the pdf should be saved.
By default, the path is an empty string so the pdf is be saved in the current working directory.}

\item{pdf_name}{string. Specifies the name of the pdf.
By default, the name of the pdf is \code{CPASS_SUBJECT_X.pdf}, where \code{X} is replaced by the subject unique identifier.}

\item{color_max_score}{string specifying the color of a score of 6 (the maximal score) reported by a subject. Any standard color format specification is accepted, i.e. one of the R built-in color names (e.g. "tomato" (default); type \code{colors()} to see the names of all R built-in colors), an RGB hex code (e.g. "#AA2199") or a color specified via one of the color/palette functions (e.g. hsv(0.1,0.9,0.9))}

\item{color_summary}{string. Either \code{"complementary"} (default) or \code{"rainbow"} specifying the type of color scheme for the ITEM items in the diagnosis summary. If \code{"complementary"}, the colors are chosen complementary within a domain; if \code{"rainbow"}, the ITEM colors are different for each item and chosen from a rainbow palette.}
}
\value{
a ggplot object if \code{save_as_pdf = FALSE}. The same object is returned invisibly by default (\code{save_as_pdf = TRUE}).
}
\description{
By default, this function saves the visualization in a pdf. The pdf heigth scales with the number of cycles reported for the subject.
}
\examples{

library(magrittr)
library(dplyr)

data(PMDD_data)
input = PMDD_data \%>\% filter(SUBJECT == 2) \%>\%  as_cpass_data(., sep_event = "menses")
p = plot_subject_data_and_diagnosis(data = input)
p
}
\keyword{C-PASS}
\keyword{CPASS}
\keyword{MRMD}
\keyword{PMDD}
